requirements.txt
python-can
cantools
